[GTK] Invalidate the ProcessLauncher when the process is terminated before it has finished launching https://bugs.webkit.org/show_bug.cgi?id=117412
Reviewed by Anders Carlsson.

* UIProcess/Launcher/gtk/ProcessLauncherGtk.cpp:
(WebKit::ProcessLauncher::terminateProcess): If process is still
launching just invalidate the launcher. Reset the process
identifier after killing the process.

Canonical link: https://commits.webkit.org/135714@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@151508 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebKit2/UIProcess/Launcher/gtk/ProcessLauncherGtk.cpp
@@ -105,11 +105,17 @@ void ProcessLauncher::launchProcess()
 }
 
 void ProcessLauncher::terminateProcess()
-{   
+{
+    if (m_isLaunching) {
+        invalidate();
+        return;
+    }
+
     if (!m_processIdentifier)
         return;
 
     kill(m_processIdentifier, SIGKILL);
+    m_processIdentifier = 0;
 }
 
 void ProcessLauncher::platformInvalidate()
